Added Support for Multiple Portraits | Fixed Text not Appearing | Fixed Options Menu Crashing
For Multiple portraits they have to be called like "portrait_". So "portrait_mad". How the game reads this is through the dialogue txt. In the dialogue txt, put this: ":dad:What did you just say to me!?:angry: | Obviously the angry being the expression. SO check for an image called "portrait_angry". The game would most likely crash cause I'm too lazy to add a quickfix for your mistakes.